moved change_networks function and other related functions to a separate file wifi.py
